Adding a gui
Making gui work well enough for now.

Adding scroll panel

Adding buttons to the output so user can save or copy the output.

formatting

Changing the name of BookCommand.java to BookSignatureFinder.java. Making it take actual arguments rather than a String array and using a callback to get the output.

Also making the buttons be enabled or disabled when appropriate.

Making it so the tool can run pdktk itself

Allowing user to specify the number of options to give.

Refactoring to allow user to specify the min and max signature sizes

making the gui look right for min and max signature inputs. Also tidying up any unused things a bit

Making the sliders look and work right